feat: remove host from db and create player-presence component
