+++ client/src/app/room/[code]/page.tsx
@@ -4,6 +4,7 @@ import ConnectionStatus from "@ai/components/connection-status";
 import { useStore } from "@ai/utils/store";
 import StoreInitializer from "@ai/utils/store-initializer";
 import { getRoomInfo } from "@ai/app/actions";
+import PlayerPresence from "@ai/utils/player-presence";
 
 export default async function Room({ params }: { params: { code: string } }) {
   const roomInfo = await getRoomInfo(params.code);
@@ -15,6 +16,7 @@ export default async function Room({ params }: { params: { code: string } }) {
   return (
     <main className="flex min-h-screen flex-col justify-center">
       <StoreInitializer room={room} players={players} />
+      <PlayerPresence code={room.code} />
       <section className="container mx-auto px-4">
         <p className="mb-2 text-center font-space text-xl">Your Room Link is</p>
         <RoomLink code={params.code} />
