+++ client/src/types/api.type.ts
@@ -35,15 +35,10 @@ export type GetRoomResponse = {
 export type RoomInfo = {
   code: string;
   created_at: string | null;
-  players: Player[];
+  players: User[];
 };
 
 // Supabase
 
 export type User = Database["public"]["Tables"]["users"]["Row"];
 export type Room = Database["public"]["Tables"]["rooms"]["Row"];
-
-export type Player = {
-  host: boolean;
-  users: User[];
-};
